requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
streamlit>=1.32.0
requests>=2.31.0
